added confirmation before delete

+++ src/http-constants.js
@@ -47,13 +47,16 @@ export const getTextRoutine = () => new Promise(
 
 export const deleteTextRoutine = (text_id) => new Promise(
 		(resolve, reject) => {
-			axios({url: baseURL+"text/"+text_id, headers:{"Authorization" : "Bearer " + localStorage.getItem('user-token') }, method: 'DELETE' })
-			.then(resp => {
-				resolve(resp)
-			})
-			.catch(err => {
-				reject(err);
-			})
+			let r = confirm("Are you sure?");
+			if(r){
+				axios({url: baseURL+"text/"+text_id, headers:{"Authorization" : "Bearer " + localStorage.getItem('user-token') }, method: 'DELETE' })
+				.then(resp => {
+					resolve(resp)
+				})
+				.catch(err => {
+					reject(err);
+				})
+			}
 		}
 	)
